+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/RandomTrackSelection.java
@@ -18,6 +18,8 @@ package com.google.android.exoplayer2.trackselection;
 import android.os.SystemClock;
 import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
+import com.google.android.exoplayer2.Timeline;
+import com.google.android.exoplayer2.source.MediaSource.MediaPeriodId;
 import com.google.android.exoplayer2.source.TrackGroup;
 import com.google.android.exoplayer2.source.chunk.MediaChunk;
 import com.google.android.exoplayer2.source.chunk.MediaChunkIterator;
@@ -51,7 +53,10 @@ public final class RandomTrackSelection extends BaseTrackSelection {
 
     @Override
     public @NullableType TrackSelection[] createTrackSelections(
-        @NullableType Definition[] definitions, BandwidthMeter bandwidthMeter) {
+        @NullableType Definition[] definitions,
+        BandwidthMeter bandwidthMeter,
+        MediaPeriodId mediaPeriodId,
+        Timeline timeline) {
       return TrackSelectionUtil.createTrackSelectionsForDefinitions(
           definitions,
           definition -> new RandomTrackSelection(definition.group, definition.tracks, random));